[Refactor] move state listener out of GlobalStateMgr, and rename it to state exchange executor
